Fix: C++ compatibility

+++ tests/segmentation.c
@@ -19,41 +19,41 @@ static void test_basic_segmentation(void) {
         state.index = 0; \
         index = 0; \
 
-    ATT_ASSERT(mjb_segmentation("", 0, MJB_ENCODING_UTF_8, &state), MJB_BT_NOT_SET, "Empty string")
+    ATT_ASSERT((uint8_t)mjb_segmentation("", 0, MJB_ENCODING_UTF_8, &state), (uint8_t)MJB_BT_NOT_SET, "Empty string")
 
     MJB_TEST_S
     mjb_break_type expected_a[] = { MJB_BT_ALLOWED };
 
     while((bt = mjb_segmentation("A", 1, MJB_ENCODING_UTF_8, &state)) != MJB_BT_NOT_SET) {
-        ATT_ASSERT(bt, expected_a[index++], "A test")
+        ATT_ASSERT((uint8_t)bt, (uint8_t)expected_a[index++], "A test")
     }
 
     MJB_TEST_S
     mjb_break_type expected_ab[] = { MJB_BT_ALLOWED, MJB_BT_ALLOWED };
 
     while((bt = mjb_segmentation("AB", 2, MJB_ENCODING_UTF_8, &state)) != MJB_BT_NOT_SET) {
-        ATT_ASSERT(bt, expected_ab[index++], "AB test")
+        ATT_ASSERT((uint8_t)bt, (uint8_t)expected_ab[index++], "AB test")
     }
     ATT_ASSERT(index, 2, "AB test break index")
 
     MJB_TEST_S
     mjb_break_type expected_abc[] = { MJB_BT_ALLOWED, MJB_BT_ALLOWED, MJB_BT_ALLOWED };
     while((bt = mjb_segmentation("ABC", 3, MJB_ENCODING_UTF_8, &state)) != MJB_BT_NOT_SET) {
-        ATT_ASSERT(bt, expected_abc[index++], "AB test")
+        ATT_ASSERT((uint8_t)bt, (uint8_t)expected_abc[index++], "AB test")
     }
     ATT_ASSERT(index, 3, "ABC test break index")
 
     MJB_TEST_S
     mjb_break_type expected_brnl[] = { MJB_BT_ALLOWED, MJB_BT_NO_BREAK, MJB_BT_ALLOWED, MJB_BT_ALLOWED };
     while((bt = mjb_segmentation("A\r\nB", 4, MJB_ENCODING_UTF_8, &state)) != MJB_BT_NOT_SET) {
-        ATT_ASSERT(bt, expected_brnl[index++], "A\\r\\nB test")
+        ATT_ASSERT((uint8_t)bt, (uint8_t)expected_brnl[index++], "A\\r\\nB test")
     }
     ATT_ASSERT(index, 4, "A\\r\\nB test break index")
 
     MJB_TEST_S
     mjb_break_type expected_itit[] = { MJB_BT_NO_BREAK, MJB_BT_ALLOWED, MJB_BT_NO_BREAK, MJB_BT_ALLOWED };
     while((bt = mjb_segmentation("🇮🇹🇮🇹", 16, MJB_ENCODING_UTF_8, &state)) != MJB_BT_NOT_SET) {
-        ATT_ASSERT(bt, expected_itit[index++], "ITIT test")
+        ATT_ASSERT((uint8_t)bt, (uint8_t)expected_itit[index++], "ITIT test")
     }
     ATT_ASSERT(index, 4, "ITIT test break index")
 
@@ -64,7 +64,7 @@ void *test_segmentation(void *arg) {
     test_basic_segmentation();
 
     char line[2048]; // 16384, see line #19335
-    char generated_input[1024]; // String to be used by mjb_break_line
+    char generated_input[1024]; // String to be used by mjb_segmentation
     mjb_break_type expected_types[1024];
     char test_name[256];
     unsigned int current_line = 1;
@@ -161,7 +161,7 @@ void *test_segmentation(void *arg) {
 
         while((bt = mjb_segmentation(generated_input, generated_index, MJB_ENCODING_UTF_8, &state)) != MJB_BT_NOT_SET) {
             snprintf(test_name, 256, "Index %zu", index);
-            ATT_ASSERT(bt, expected_types[index++], test_name)
+            ATT_ASSERT((uint8_t)bt, (uint8_t)expected_types[index++], test_name)
         }
 
         free(tofree);
